# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OPENCUT API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Type of wallet
 */
export type WalletInputWalletType = typeof WalletInputWalletType[keyof typeof WalletInputWalletType];


export const WalletInputWalletType = {
  metamask: 'metamask',
  phantom: 'phantom',
} as const;

export interface WalletInput {
  /** Blockchain wallet address */
  walletAddress: string;
  /** Type of wallet */
  walletType: WalletInputWalletType;
  /**
     * Optional display name
     * @nullable
     */
  username?: string | null;
}

export interface UserUpdate {
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
}

export type UserWalletType = typeof UserWalletType[keyof typeof UserWalletType];


export const UserWalletType = {
  metamask: 'metamask',
  phantom: 'phantom',
} as const;

export interface User {
  id: number;
  walletAddress: string;
  walletType: UserWalletType;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  joinedAt: string;
  lastSeenAt?: string;
}

export interface UserStats {
  projectsCount: number;
  exportsCount: number;
  totalDurationSeconds: number;
  memberSince: string;
}

export interface LeaderboardEntry {
  rank: number;
  walletAddress: string;
  walletType: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  exportsCount: number;
}

export interface ErrorResponse {
  error: string;
}