# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OPENCUT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  walletAddress: string;
  walletType: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  exportsCount: number;
}