# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OPENCUT API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserWalletType } from './userWalletType';

export interface User {
  id: number;
  walletAddress: string;
  walletType: UserWalletType;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  joinedAt: Date;
  lastSeenAt?: Date;
}