# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OPENCUT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  projectsCount: number;
  exportsCount: number;
  totalDurationSeconds: number;
  memberSince: Date;
}