# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OPENCUT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
}