# lib/api-zod/src/generated/types/walletInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OPENCUT API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WalletInputWalletType } from './walletInputWalletType';

export interface WalletInput {
  /** Blockchain wallet address */
  walletAddress: string;
  /** Type of wallet */
  walletType: WalletInputWalletType;
  /**
     * Optional display name
     * @nullable
     */
  username?: string | null;
}